requirements.txt
# Core dependencies
langchain
langchain-community
langchain-google-genai
google-generativeai

streamlit
python-dotenv
PyPDF2

